refactor: extract cancellation scheduling into dedicated classes
Address the architectural review on PR #10:
- Extract ScheduledAction out of MockClock into its own internal class so
  it can be shared and subclassed.
- Add an internal ScheduledCancellationCollection (owned by a MockClock)
  that tracks at most one pending cancellation per source and exposes
  AddOrReplace(source, timeout). Replacement is atomic under one lock, so
  the latest deadline wins with no lost update under concurrent calls.
- Add a ScheduledCancellation : ScheduledAction (a private nested class of
  the collection) that cancels its source and then removes its own entry
  from the collection by key.

MockClock.CancelAfter now just validates (null, range, disposed probe,
already-cancelled) and delegates to the collection. Behaviour is unchanged;
all 64 tests pass on net48/net8.0/net10.0.
